Update README with processing module documentation
